Modularize some helper functions in Destinations Viewer per Jason's suggestion for readability. Could not modularize renderCities because it needs to be in scope. This will do for now
